requirements.txt
asgiref==3.4.1
atomicwrites==1.4.0
attrs==21.2.0
autopep8==1.5.7
colorama==0.4.4
Django==3.2.5
django-phonenumber-field==5.2.0
iniconfig==1.1.1
packaging==21.0
phonenumberslite==8.12.26
Pillow==8.3.0
pluggy==0.13.1
py==1.10.0
pycodestyle==2.7.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.4.0
python-dotenv==0.18.0
pytz==2021.1
sqlparse==0.4.1
toml==0.10.2
